requirements.txt
# Core dependencies
asgiref==3.7.2
Django==4.2.20
python-dotenv==1.0.0
gunicorn==23.0.0
whitenoise==6.6.0

# Database
psycopg2-binary==2.9.9  # For PostgreSQL (commented out for now)

# Storage
boto3==1.25.5
botocore==1.28.5
django-storages==1.13.1
s3transfer==0.6.0

# Django packages
django-allow-cidr==0.5.0
django-crispy-forms==1.14.0
django-mathfilters==1.0.0
django-utils-six==2.0

# Utilities
jmespath==1.0.1
packaging==21.3
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.12.2
tzdata==2022.5
urllib3==1.26.12

# Uncomment if needed
# Pillow==10.1.0  # For image processing